+++ Assets/GothicVR/Scripts/Creator/Meshes/PolyStripMeshCreator.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using GVR.Caches;
 using GVR.Creator.Meshes;
@@ -7,6 +8,8 @@ using UnityEngine;
 
 public class PolyStripMeshCreator : AbstractMeshCreator
 {
+
+    [Obsolete("Use MeshFactory and *MeshBuilder instead.")]
     public void CreatePolyStrip(GameObject go, int numberOfSegments, Vector3 startPoint, Vector3 endPoint)
     {
         var material = new Material(Constants.ShaderThunder);
