+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: b67d4aa7d8a24910ab244f7e4a0aa1fa
+timeCreated: 1711196533
