+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/AbstractMeshBuilder.cs.meta
