+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/BarrierMeshBuilder.cs.meta
