+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/NpcHeadMeshBuilder.cs.meta
