+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/NpcMeshBuilder.cs.meta
