+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/NpcWeaponMeshBuilder.cs.meta
