+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/PolyStripMeshBuilder.cs.meta
