+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/Textures/WorldTextureArrayBuilder.cs.meta
