+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/VobDecalMeshBuilder.cs.meta
