+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/VobMeshBuilder.cs
@@ -4,7 +4,7 @@ using GVR.Globals;
 using Unity.VisualScripting;
 using UnityEngine;
 
-namespace GVR.Creator.Meshes.V2
+namespace GVR.Creator.Meshes.V2.Builder
 {
     public class VobMeshBuilder : AbstractMeshBuilder
     {
