+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/VobMeshBuilder.cs.meta
