+++ Assets/GothicVR/Scripts/Creator/Meshes/V2/Builder/WorldMeshBuilder.cs
@@ -10,7 +10,7 @@ using UnityEngine;
 using ZenKit;
 using Mesh = UnityEngine.Mesh;
 
-namespace GVR.Creator.Meshes.V2
+namespace GVR.Creator.Meshes.V2.Builder
 {
     public class WorldMeshBuilder : AbstractMeshBuilder
     {
