+++ Assets/GothicVR/Scripts/Creator/Meshes/VobMeshCreator.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using GVR.Caches;
 using GVR.Extensions;
@@ -12,6 +13,8 @@ using Material = UnityEngine.Material;
 
 namespace GVR.Creator.Meshes
 {
+
+    [Obsolete("Use MeshFactory and *MeshBuilder instead.")]
     public class VobMeshCreator : AbstractMeshCreator
     {
         public GameObject CreateVob(string objectName, IMultiResolutionMesh mrm, Vector3 position,
